+++ platform-core-module/src/test/java/com/softicar/platform/core/module/program/execution/queued/unqueue/ProgramUnqueuerTest.java
@@ -2,13 +2,12 @@ package com.softicar.platform.core.module.program.execution.queued.unqueue;
 
 import com.softicar.platform.common.date.DayTime;
 import com.softicar.platform.core.module.program.AGProgram;
+import com.softicar.platform.core.module.program.AbstractProgramTest;
 import com.softicar.platform.core.module.program.unqueue.ProgramUnqueuer;
-import com.softicar.platform.core.module.user.AGUser;
-import com.softicar.platform.db.runtime.test.AbstractDbTest;
 import java.util.UUID;
 import org.junit.Test;
 
-public class ProgramUnqueuerTest extends AbstractDbTest {
+public class ProgramUnqueuerTest extends AbstractProgramTest {
 
 	private static final UUID SOME_UUID = UUID.fromString("28818f64-369a-412c-a27e-c6697800a600");
 	private final AGProgram program;
@@ -39,23 +38,12 @@ public class ProgramUnqueuerTest extends AbstractDbTest {
 
 		program//
 			.setQueuedAt(DayTime.now())
-			.setQueuedBy(insertUser())
+			.setQueuedBy(user)
 			.save();
 
 		boolean removed = new ProgramUnqueuer(program).removeFromQueue();
 
 		assertTrue(removed);
 		assertNull(program.getQueuedAt());
-	}
-
-	private AGUser insertUser() {
-
-		return new AGUser()//
-			.setActive(false)
-			.setLoginName("john.doe")
-			.setFirstName("John")
-			.setLastName("Doe")
-			.setEmailAddress("john.doe@example.com")
-			.save();
 	}
 }
